\name{myvariant}
\alias{myvariant}
\docType{package}
\title{
Access MyVariant.info variant annotation services

}
\description{
MyVariant.Info provides REST web services to query/retrieve variant annotations. myvariant is an easy-to-use R wrapper to access MyVariant.info services.
}
\details{
\tabular{ll}{
Package: \tab myvariant\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2014-12-18\cr
License: \tab Artistic-2.0\cr
Depends: \tab httr jsonlite Hmisc\cr
}
}
\author{
Adam Mark

Maintainer: Adam Mark <adammark@scripps.edu>
 
}
\references{
https://github.com/SuLab/myvariant.info/wiki
}

\keyword{ package }

