\name{MyVariant}
\alias{MyVariant}

\title{
MyVariant
}
\description{
Construct a MyVariant object.
}

\usage{
MyVariant(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{...}{
See help page for MyVariant-class
}
}

\value{MyVariant object}

\examples{
MyVariant()
}

